requirements.txt
django==1.4.5
django-mptt==0.5.2